requirements.txt
albumentations==1.4.4
einops==0.8.0
Flask==3.0.3
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
PyYAML==6.0
PyYAML==6.0.1
scipy==1.10.1
tensorflow==2.16.1
tensorflow_probability==0.24.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
tqdm==4.65.0
transformers==4.37.2
